REFACTOR:#244 Comment 쿼리 개선 및 버그 해결
