+++ react-app/src/styles/index.css
@@ -69,4 +69,17 @@
   .bg-orange-gradient {
     background: linear-gradient(180deg, var(--color-orange) 60%, #ffc18f 100%);
   }
+
+  .custom-scrollbar::-webkit-scrollbar {
+    width: 6px;
+  }
+
+  .custom-scrollbar::-webkit-scrollbar-track {
+    background: transparent;
+  }
+
+  .custom-scrollbar::-webkit-scrollbar-thumb {
+    background-color: rgba(0, 0, 0, 0.15);
+    border-radius: 4px;
+  }
 }
